+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Inspect a value set for `translator` modifier [#901](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/901)
 - Inspect a value set for `cellDecorator` modifier [#902](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/902)
 - Inspect a value set for `mode` modifier [#903](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/903)
+- Inspect a value set for modifier of boolean type [#904](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/904)
 
 ### `FlexibleSearch` enhancements
 - Improved `words case` and `table alias` actions performance [#874](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/874)